(maint) Acceptance - handle test failure in the event of a log entry not appearing
We use beaker's retry_on to patiently allow some time for a log entry to appear. In the event of the log entry failing to appear; we should have the test fail rather than get into an error state.
This commit adds a method to the SSL config test to rescue an exception from retry_on and call an assertion statement that will give us an explicit test failure.